Plot refactoring
- Split inserting layout and creating live spec into separate methods.
- Fixed typo in layout file name.
- Updated type definition.
